requirements.txt
python-telegram-bot[job-queue]
anthropic
python-dotenv
sqlalchemy
alembic
psycopg2-binary
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
